Feature: debounce storage updates
Feature:
 - debounce updates
Refactor:
- move local storage code to new file and fix page title
